Fix problems and testing requests and response.

+++ src/main/java/by/com/gpsolutions/hotelmicro/controller/HistogramController.java
@@ -16,7 +16,7 @@ public class HistogramController {
     }
 
     @GetMapping("/{param}")
-    public Map<String, Integer> getHistogram(@PathVariable String param) {
+    public Map<String, Long> getHistogram(@PathVariable String param) {
         return hotelService.getHistogram(param);
     }
